Version 0.5.0: *** Major update ***
The IPlayable system has been overhauled and simplified in this version. This is a system-level change, and it's a shame to have it come after the Open Source Flash Development book has been published, but such is life. As GoASAP has gotten more use it's become clear that PlayableBase was too primary -- often subclassing is not a good option unless you're working from scratch. Keeping the PlayStates constants in their own class loosens things up quite a bit and provides a lot more freedom.

1. The IPlayableBase interface has been removed from the GoASAP core. The state property was moved into the IPlayable interface, and the playableID property is now treated as an optional convention.

2. The play-state constants (PLAYING, etc.) have been moved into a new class called PlayStates.

3. The class PlayableBase is now treated as entirely optional

This update has a ripple effect, many of the example/test classes and utilities have been updated to use the new PlayStates. Also includes a small improvement in Sequence.skipTo() and some documentation updates.

[Special thanks to Tim Knip, Donovan Adams and Jack Doyle for their input on this rev!]